domains/automation-core/automation-trace-service/requirements.txt
# Automation Trace Service Dependencies
fastapi>=0.123.0,<0.124.0
uvicorn[standard]>=0.32.0,<0.33.0
aiohttp>=3.13.2,<4.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
influxdb3-python>=0.3.0,<1.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
httpx>=0.28.1,<0.29.0

# HomeIQ shared libraries (installed from libs/ in Docker)
homeiq-observability>=1.0.0
